Moved signals to models.py
