+++ accounts/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 from django.contrib.auth.models import User
+from django.db.models.signals import post_save
+from django.dispatch import receiver
 # TODO: add email verification - issue<41>
 
 
@@ -18,3 +20,14 @@ class UserAccount(models.Model):
 
     def __str__(self):
         return str(self.user)
+
+
+@receiver(post_save, sender=User)
+def create_user_account(sender, instance, created, **kwargs):
+    if created:
+        UserAccount.objects.create(user=instance)
+
+
+@receiver(post_save, sender=User)
+def save_user_account(sender, instance, **kwargs):
+    instance.useraccount.save()
